ingestion/requirements.txt
pandas
pyarrow
python-chess
zstandard